+++ CHANGELOG.md
@@ -8,8 +8,8 @@ CHANGELOG
 3.7.0
 -----
 
-* remove pattern added from the Expression Violation message.
 * Added `@SerializedName` annotation support and name converters when using Symfony >= 4.2.
+* remove pattern added from the Expression Violation message.
 
 3.3.0
 -----