+++ CHANGELOG.md
@@ -5,8 +5,9 @@ CHANGELOG
 ------------------
 
 * Added `@SerializedName` annotation support and name converters when using Symfony >= 4.2.
-* remove pattern added from the Expression Violation message. 
-
+* Removed pattern added from the Expression Violation message.
+* Added FOSRestBundle 3.x support
+* Added `@SWG` annotations support at methods level in models
 3.3.0
 -----