+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 CHANGELOG
 =========
 
-3.7.0 (unreleased)
-------------------
+3.7.0
+-----
 
 * Added `@SerializedName` annotation support and name converters when using Symfony >= 4.2.
 * Removed pattern added from the Expression Violation message.
 * Added FOSRestBundle 3.x support
 * Added `@SWG` annotations support at methods level in models
+
 3.3.0
 -----